Add neo-vim session reload functionality
- Add version flag
