+++ alacritty_colorscheme/cli.py
@@ -1,28 +1,17 @@
 #!/usr/bin/env python3
 
 from tap import Tap
-from re import match
 from tempfile import NamedTemporaryFile
 from typing import List, Optional, Literal, cast
-from os import listdir, unlink
-from os.path import expanduser, isfile, join, splitext
-from shutil import copyfile
-from ruamel.yaml import YAML
-from ruamel.yaml.error import CommentMark
-from ruamel.yaml.tokens import CommentToken
+from os import listdir
+from os.path import expanduser, isfile, join
 
-yaml = YAML()
-yaml.indent(mapping=2, sequence=4, offset=2)
-
-config_path = join('~', '.config/alacritty/alacritty.yml')
-colorscheme_dir = join('~', '.config/alacritty/colors/')
+from . import __version__
+from .colorscheme import get_applied_colorscheme, get_applicable_colorscheme, replace_colorscheme
 
 
-class ApplyParser(Tap):
-    colorscheme: str
-
-    def configure(self) -> None:
-        self.add_argument('colorscheme')
+class StatusParser(Tap):
+    pass
 
 
 # TODO: filter dark and light backgrounds
@@ -30,8 +19,11 @@ class ListParser(Tap):
     pass
 
 
-class StatusParser(Tap):
-    pass
+class ApplyParser(Tap):
+    colorscheme: str
+
+    def configure(self) -> None:
+        self.add_argument('colorscheme')
 
 
 class ToggleParser(Tap):
@@ -42,10 +34,15 @@ class ToggleParser(Tap):
         self.add_argument('colorschemes')
 
 
-class SimpleArgumentParser(Tap):
+config_path = join('~', '.config/alacritty/alacritty.yml')
+colorscheme_dir = join('~', '.config/alacritty/colors/')
+
+
+class ArgumentParser(Tap):
     config_file: str = config_path  # Path to alacritty configuration file
     colorscheme_dir: str = colorscheme_dir  # Path to colorscheme directory
     base16_vim: bool = False  # Support base16-vim. Generates .vimrc_background file at home directory
+    # version: str  # Version
 
     def configure(self) -> None:
         self.add_subparsers(help='sub-command help', dest="_subparser_name")
@@ -65,15 +62,20 @@ class SimpleArgumentParser(Tap):
         self.add_argument('-V',
                           '--base16_vim')
 
+        self.add_argument('-v',
+                          '--version',
+                          action='version',
+                          version='%(prog)s {version}'.format(version=__version__))
 
-# NOTE: adding this to SimpleArgumentParser will add a new argument.
-# Creating this class for casting purpose only
-class HackArgumentParser(SimpleArgumentParser):
+
+# NOTE: adding '_subparser_naem' to ArgumentParser will add a new argument.
+# So, creating this class for type casting purpose only
+class HackArgumentParser(ArgumentParser):
     _subparser_name: Literal['list', 'status', 'toggle', 'apply']
 
 
-def parse_args() -> SimpleArgumentParser:
-    parser = SimpleArgumentParser(
+def parse_args() -> ArgumentParser:
+    parser = ArgumentParser(
         "alacritty-colorscheme",
         description="Change colorscheme of alacritty with ease."
     )
@@ -85,143 +87,10 @@ def parse_args() -> SimpleArgumentParser:
 def get_files_in_directory(path: str) -> Optional[List[str]]:
     expanded_path = expanduser(path)
     try:
-        onlyfiles = [f for f in (listdir(expanded_path))
+        onlyfiles = [f for f in listdir(expanded_path)
                      if isfile(join(expanded_path, f))]
         return sorted(onlyfiles)
     except OSError:
-        return None
-
-
-def template_vimrc_background(colorscheme: str) -> str:
-    command = (
-        f"if !exists('g:colors_name') || g:colors_name != '{colorscheme}'\n"
-        f"  colorscheme {colorscheme}\n"
-        "endif")
-    return command
-
-
-# NOTE: function to identify if a given `yaml.Comment` internally has
-# at least one comment
-def has_comment_token(colors_comment: List) -> bool:
-    if not colors_comment or len(colors_comment) < 2:
-        return False
-
-    comment_tokens = colors_comment[1]
-    return comment_tokens and len(comment_tokens) >= 1
-
-
-def get_applied_colorscheme(config_path: str) -> Optional[str]:
-    with open(expanduser(config_path), 'r') as config_file:
-        config_yaml = yaml.load(config_file)
-
-    if not has_comment_token(config_yaml['colors'].ca.comment):
-        return None
-
-    comment_match = match(
-        r"#\s*COLORSCHEME:\s*(.*)\s*\n",
-        config_yaml['colors'].ca.comment[1][0].value,
-    )
-
-    if not comment_match:
-        return None
-
-    comment_groups = comment_match.groups()
-    return comment_groups[0]
-
-
-def replace_colorscheme(
-    colors_path: str,
-    config_path: str,
-    colorscheme: str,
-    base16_vim: bool,
-) -> None:
-    try:
-        with open(expanduser(config_path), 'r') as config_file:
-            config_yaml = yaml.load(config_file)
-    except OSError:
-        print(f'Could not find a valid alacritty config file: {config_path}')
-        return
-
-    try:
-        with open(expanduser(colors_path), 'r') as color_file:
-            colors_yaml = yaml.load(color_file)
-    except OSError:
-        print(f'Could not find a valid alacritty colorscheme file: {colors_path}')
-        return
-
-    try:
-        # NOTE: update method doesn't read the first comment
-        config_yaml['colors'].update(colors_yaml['colors'])
-    except KeyError:
-        config_yaml['colors'] = colors_yaml['colors']
-
-    new_comment_token = CommentToken(
-        f'# COLORSCHEME: {colorscheme}\n',
-        CommentMark(2),
-        None,
-    )
-
-    if has_comment_token(config_yaml['colors'].ca.comment):
-        # removing all comments for colors in config_file
-        while len(config_yaml['colors'].ca.comment[1]) > 0:
-            config_yaml['colors'].ca.comment[1].pop()
-
-        # adding current colorscheme name in comment
-        config_yaml['colors'].ca.comment[1].append(new_comment_token)
-    else:
-        # adding current colorscheme name in comment
-        config_yaml['colors'].ca.comment = [None, [new_comment_token]]
-
-    # adding all comments for colors from colors_file
-    if has_comment_token(colors_yaml['colors'].ca.comment):
-        config_yaml['colors'].ca.comment[1].extend(
-            colors_yaml['colors'].ca.comment[1]
-        )
-
-    try:
-        with NamedTemporaryFile(delete=False) as tmp_file:
-            # NOTE: not directly writing to config_file as it causes
-            # multiple reload during write
-            tmp_file_path = tmp_file.name
-            yaml.dump(config_yaml, tmp_file)
-            copyfile(tmp_file_path, expanduser(config_path))
-            unlink(tmp_file_path)
-    except OSError:
-        print(f'Could not modify alacritty config file: {config_path}')
-        return
-
-    if base16_vim:
-        vimrc_background_path = join('~', '.vimrc_background')
-        try:
-            with open(expanduser(vimrc_background_path), 'w') as vimrc_background_file:
-                colorscheme_no_extension = splitext(colorscheme)[0]
-                vimrc_background_content = template_vimrc_background(
-                    colorscheme_no_extension
-                )
-                vimrc_background_file.write(vimrc_background_content)
-        except OSError:
-            print(f'Could not save file: {vimrc_background_path}')
-            return
-
-
-def get_applicable_colorscheme(
-    colorschemes: List[str],
-    colorscheme: Optional[str],
-    reverse_toggle: bool
-) -> Optional[str]:
-    if colorscheme is None:
-        index = 0
-    else:
-        try:
-            original_index = colorschemes.index(colorscheme)
-            diff = - 1 if reverse_toggle else 1
-            index = (original_index + diff) % len(colorschemes)
-        except ValueError:
-            index = 0
-
-    try:
-        return colorschemes[index]
-    except IndexError:
         return None
